+++ backend/src/main/java/com/woowacourse/gongseek/member/application/MemberService.java
@@ -8,10 +8,10 @@ import com.woowacourse.gongseek.comment.domain.repository.CommentRepository;
 import com.woowacourse.gongseek.member.domain.Member;
 import com.woowacourse.gongseek.member.domain.repository.MemberRepository;
 import com.woowacourse.gongseek.member.exception.MemberNotFoundException;
-import com.woowacourse.gongseek.member.presentation.MyPageArticleResponse;
-import com.woowacourse.gongseek.member.presentation.MyPageArticlesResponse;
-import com.woowacourse.gongseek.member.presentation.MyPageCommentResponse;
-import com.woowacourse.gongseek.member.presentation.MyPageCommentsResponse;
+import com.woowacourse.gongseek.member.presentation.dto.MyPageArticleResponse;
+import com.woowacourse.gongseek.member.presentation.dto.MyPageArticlesResponse;
+import com.woowacourse.gongseek.member.presentation.dto.MyPageCommentResponse;
+import com.woowacourse.gongseek.member.presentation.dto.MyPageCommentsResponse;
 import com.woowacourse.gongseek.member.presentation.dto.MemberDto;
 import java.util.List;
 import java.util.stream.Collectors;
