+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/MemberAcceptanceTest.java
@@ -9,8 +9,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.woowacourse.gongseek.article.presentation.dto.ArticleIdResponse;
 import com.woowacourse.gongseek.auth.presentation.dto.TokenResponse;
-import com.woowacourse.gongseek.member.presentation.MyPageArticlesResponse;
-import com.woowacourse.gongseek.member.presentation.MyPageCommentsResponse;
+import com.woowacourse.gongseek.member.presentation.dto.MyPageArticlesResponse;
+import com.woowacourse.gongseek.member.presentation.dto.MyPageCommentsResponse;
 import com.woowacourse.gongseek.member.presentation.dto.MemberDto;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
