+++ backend/src/test/java/com/woowacourse/gongseek/member/application/MemberServiceTest.java
@@ -15,8 +15,8 @@ import com.woowacourse.gongseek.commons.DatabaseCleaner;
 import com.woowacourse.gongseek.member.domain.Member;
 import com.woowacourse.gongseek.member.domain.repository.MemberRepository;
 import com.woowacourse.gongseek.member.exception.MemberNotFoundException;
-import com.woowacourse.gongseek.member.presentation.MyPageArticlesResponse;
-import com.woowacourse.gongseek.member.presentation.MyPageCommentsResponse;
+import com.woowacourse.gongseek.member.presentation.dto.MyPageArticlesResponse;
+import com.woowacourse.gongseek.member.presentation.dto.MyPageCommentsResponse;
 import com.woowacourse.gongseek.member.presentation.dto.MemberDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
